解决前台的 Float 字段问题 Float 字段改名为 Double

+++ src/ObjectServer.Client.Agos/Windows/FormView/FieldControlFactory.cs
@@ -31,7 +31,7 @@ namespace ObjectServer.Client.Agos.Windows.FormView
                 { "reference", typeof(ReferenceFieldControl) },
                 { "datetime", typeof(DateTimeFieldControl) },
                 { "date", typeof(DateFieldControl) },
-                { "float8", typeof(FloatFieldControl) },
+                { "double", typeof(FloatFieldControl) },
                 { "decimal", typeof(DecimalFieldControl) },
             };
